flask-app/requirements.txt
Flask==2.2.5
pymongo==3.12.0
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
flask-mongoengine-3==1.0.8
Werkzeug>=3.0.0
PyJWT==2.7.0
Flask-JWT-Extended==4.4.0
redis==5.2.0